Fixed case sensative uri matching in transport options
