+++ server/src/main/java/org/kaazing/gateway/server/context/resolve/DefaultOptionsContext.java
@@ -17,6 +17,7 @@ package org.kaazing.gateway.server.context.resolve;
 import static java.lang.String.format;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.kaazing.gateway.resource.address.uri.URIUtils.getCanonicalizedURI;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -97,6 +98,7 @@ class DefaultOptionsContext {
                 throw new IllegalArgumentException(format(
                         "%s must contain an absolute URI, not \"%s\"", transportKey, transport));
             }
+            transportURI = getCanonicalizedURI(transportURI, false);
         }
 
         return transportURI;
