+++ server/src/test/java/org/kaazing/gateway/server/config/parse/ServiceDefaultsTest.java
@@ -69,8 +69,8 @@ public class ServiceDefaultsTest {
         Assert.assertTrue("TLSv1".equals(sslProtocols[0]));
 
         Assert.assertTrue("en0".equals(connectOptionsMap.get("udp.interface")));
-        System.out.println(connectOptionsMap.get("tcp.transport"));
-        Assert.assertTrue("socks://localhost:8000".equals(connectOptionsMap.get("tcp.transport").toString().trim()));
+        final String trim = connectOptionsMap.get("tcp.transport").toString().trim();
+        Assert.assertTrue("socks://localhost:8000".equals(trim));
         Assert.assertFalse((Boolean)connectOptionsMap.get("ssl.encryptionEnabled"));
 
         Assert.assertEquals(5, connectOptionsMap.get("http[http/1.1]."+HTTP_KEEP_ALIVE_TIMEOUT_KEY));
